refactored compute_orientation
