+++ mpc_follower_wrapper/src/mpc_follower_wrapper.cpp
@@ -33,40 +33,6 @@ MPCFollowerWrapper::MPCFollowerWrapper(ros::CARMANodeHandle &nodeHandle): nh_(no
 MPCFollowerWrapper::~MPCFollowerWrapper() {
 }
 
-std::vector<geometry_msgs::Quaternion>
-compute_orientations(const cav_msgs::TrajectoryPlan::ConstPtr& tp)
-{
-  std::vector<geometry_msgs::Quaternion> out;
-
-  if (tp->trajectory_points.size() == 0) {
-      return out;
-  }
-
-  
-  lanelet::BasicLineString2d centerline;
-  for (cav_msgs::TrajectoryPlanPoint traj_point : tp->trajectory_points) {
-    lanelet::BasicPoint2d p(traj_point.x, traj_point.y);
-    centerline.push_back(p);
-  }
-
-
-  std::vector<Eigen::Vector2d> tangents = carma_wm::geometry::compute_finite_differences(centerline);
-
-  Eigen::Vector2d x_axis = { 1, 0 };
-  for (int i = 0; i < tangents.size(); i++)
-  {
-    geometry_msgs::Quaternion q;
-
-    // Derive angle by cos theta = (u . v)/(||u| * ||v||)
-    double yaw = carma_wm::geometry::safeAcos(tangents[i].dot(x_axis) / (tangents[i].norm() * x_axis.norm()));
-
-    q = tf::createQuaternionMsgFromYaw(yaw);
-    out.push_back(q);
-  }
-
-  return out;
-}
-
 void MPCFollowerWrapper::Initialize() {
 
 
@@ -102,7 +68,12 @@ void MPCFollowerWrapper::TrajectoryPlanPoseHandler(const cav_msgs::TrajectoryPla
       autoware_msgs::Lane lane;
       lane.header = tp->header;
       std::vector <autoware_msgs::Waypoint> waypoints;
-      std::vector<geometry_msgs::Quaternion> quats = compute_orientations(tp);
+      lanelet::BasicLineString2d centerline;
+      for (cav_msgs::TrajectoryPlanPoint traj_point : tp->trajectory_points) {
+        lanelet::BasicPoint2d p(traj_point.x, traj_point.y);
+        centerline.push_back(p);
+      }
+      std::vector<geometry_msgs::Quaternion> quats = carma_wm::geometry::compute_tangent_orientations(centerline);
       if (quats.size() != tp->trajectory_points.size()) {
         ROS_ERROR_STREAM("Quat list size mismatch");
       }
